requirements.txt
aiofiles==0.6.0
aiomysql==0.0.21
aiosqlite==0.16.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
colorama==0.4.4
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
iso8601==0.1.13
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.4
prompt-toolkit==3.0.8
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.0
PyMySQL==0.9.3
PyPika==0.44.1
python-dotenv==0.15.0
python-rapidjson==1.0
pytz==2020.5
PyYAML==5.3.1
six==1.15.0
starlette==0.13.6
tortoise-orm==0.16.19
typing-extensions==3.7.4.3
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
wcwidth==0.2.5
websockets==8.1
XlsxWriter==1.3.7
